move to inner classes
